Complete refactorization of the 3 search forms, stage 2
